major additions, seedbank module

+++ src/server/database/migrations/2015_10_20_153909_create_seed_months_table.php
@@ -15,7 +15,7 @@ class CreateSeedMonthsTable extends Migration
         Schema::create('seed_months', function (Blueprint $table) {
             $table->increments('id');
             $table->timestamps();
-            $table->integer('month')->unsigned();
+            $table->smallInteger('month')->unsigned();
             $table->integer('seed_id')->unsigned();
             $table->foreign('seed_id')->references('id')->on('seeds');
         });
